requirements.txt
streamlit
pyserial
pandas
tzlocal